fix: move po not found error back
