--- router/internal/persistedoperation/operationstorage/errors.go
@@ -1,14 +1,0 @@
-package operationstorage
-
-import "fmt"
-
-var PoNotFoundErr *PersistentOperationNotFoundError
-
-type PersistentOperationNotFoundError struct {
-	ClientName string
-	Sha256Hash string
-}
-
-func (e PersistentOperationNotFoundError) Error() string {
-	return fmt.Sprintf("operation %s for client %s not found", e.Sha256Hash, e.ClientName)
-}
